Removed color from Body2D. Drawing inner planetarium surface to window.
Body2D now has a color indirectly through the userObject attribute (to
reduce coupling).

getRandomColor() is back to CPlanetsGUI scope.

The Planetarium's draw() method now blits the inner surface to the
screen. This step was missing to be able to see the bodies.

Temporarily drawing the bodies as rectangles, for debugging purposes.