Created shared version of Color5 from SDL_widgets
This is due to SDL_widgets original code not providing Color5 on its header, yet there was a WinBase::draw_gradient() with a Color5 argument. Instead, this SDL_util::Color5 provides a Color5::draw_gradient() with a WinBase* argument instead of Color5 argument.

Also there are defined some constant gradients.