Spinner no longer needs manual refresh() when setting value at main().
Spinner implementation was improved, no longer requiring to call
refresh() when setting value in the initialization code. After calling
setValue(), it became scheduled to run refresh() when calling draw()
through a flag called 'needsRefresh'. If after a refresh is done
(including through calling refresh() manually), it then clears
'needsRefresh'. Note that it is still neccessary to redraw the spinner
for changes in its value become visible.

All Spinner's methods are now virtual, to help future extending of the
class.