Created custom callback to fix list selection. Fixed updating list.
Using preventRedrawOnClick flag to disable automatic draw_blit_upd()
call when clicking on the ListWin.

Created a callback callbackTxtBodiesClicked() that clicks on the list
and calls the refresh method of the scrollable pane, effectively fixing
bugs when clicking on the list when it is scrolled.

Fixed onUserEvent() callback code to fix bugs when the list was updated
through user events (body collision, creation, etc)