Added new signature for Planetarium::addCustomBody()
This new signature doesn't attempt to modify user object if it is present. If it is empty, creates a new Planetarium user object with a random color.

The signature 'void Planetarium::addCustomBody(Body2D* body, const SDL_Color& color)' will attempt to assign the given color if a user object is present. If user object is absent, it will create a new Planetarium user object, with the given color.