Created StateManager class. Moved methods related to undo to this class.
StateManager is opaque, and only seen on planetarium.cpp. This class manages the undo stack.

Behavior: the undo stack only hold copies (in Body2DClone form) for position/velocity purposes. The diffs are the actual pointers.

The previous idea with full deep copies was dropped. Planetarium::OnCollision manages the deletion of the mergers by creating new instances and updating references in the undo stack. The references the undo stack are replaced by new ones in case of merges. Body removal now only deletes the removed bodies if the undo functionality is disabled altogether. Otherwise it is saved as is on the undo stack.

Bodies are only deleted when performing StateManager::rollback or StateManager::reset().