Collided bodies no longer are passed with a copied user object.
When bodies collided, a vector with copies of them is passed to the collision listeners. But, as the original ones are deleted by the physics thread, copies of its user objects were also provided instead of the original ones. But these copy user objects were never deleted, not even by the listeners (and listeners should not need to do this). So, instead of a copy, the user object field is deleted and nullified before passing to the event queue.